[FIX] : 🔧 뷰 모듈화
