--- HanSang/View/DetailPage/TestViewController.swift
@@ -1,18 +1,0 @@
-//
-//  TestViewController.swift
-//  HanSang
-//
-//  Created by 박성원 on 2023/09/29.
-//
-
-import UIKit
-
-final class TestViewController: UIViewController {
-
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-    }
-
-}
